Begin redesign
Make the server dumb REST. Iterating on a UI in Rust is painfully slow. Instead, serve
static HTML & JS files which can be edited on the fly, and then bundled w/ the binary.

This disables a lot of functionality, lots to do to re-enable things.  But this has the
static file serving/embedding working.