Make user_known more strict.
Previously, if a user had *ever* posted an item/profile to a
server, they were "known" and could always send profile updates. This could lead to abuse. No one on the server is watching that user's content anymore, but somehow they're still allowed to maintain a profile there? Not anymore.

I should probably also check known_user before serving any content for a userID as well, so that once a user is no longer followed here, we no longer serve content for them.