WIP: NaCl in Web Workers.
To remove strain from the main browser thread, we can run some expensive crypto operations asynchronously in a web worker.

Currently does not work in Firefox or Safari. Going to try to work around that.

TBH, the main performance issue seems to have been the missing `await` fixed in the previous commit.  For fully serial code, doing the work in a web worker actually introduces a bit of overhead that slows things down. BUT, it would allow me to easily add support for a pool of webworkers, which could be handy for efficiently performing bulk operations like sync.