Add file sync support.
File sync happens after items are sync'd. This allows it to recover from temporarily unavailable files.
i.e.: If a file fails to copy, the next sync will retry it.

Right now, this attempts to sync all files, but later we can add filters for what we want to sync.

Also:
* Some extra debug logging (when enabled).
* Fix unreleased bug: store(hash) and attachment(uid,sig,name) should be UNIQUE
* During sync, the client can skip verifying signatures since we're going to send it to a server which will do it for us. Yay even faster syncs.