Fixes #51: Timestamps now default to "now".
Since the introduction of drafts, the web client would save
a post body and timestamp (string). Returning to the page, deleting the previously saved draft, and writing a new post, would result in posting a backdated post.

Since posting with the current timestamp will likely be the common case, we handle that better now. A blank timestamp input field means "now", and "now" will be updated every time you modify the title/body of your post.