Enable more efficient caching of HTTP files.
Now all of our static files are served with ETag headers, which means
that if the browser requests them again, we can just tell it that they
haven't been modified instead of re-serving them.

(We already do something similar for immutable files like Items and
attachments.)

Thanks to the rust-embed devs for implementing this feature! 🤝

Closes #43