Fix Rust 1.64 build break. :(
https://blog.rust-lang.org/2022/09/22/Rust-1.64.0.html
Rust 1.64 changes the memory layout of socket
addresses.  This broke socket2, used by actix (and FeoBlog).

Upgrading actix v3 ran into other build failures, and it
hasn't had a release in 2022, so upgraded to
Actix 4 for a working version.

Actix v4 is a new semver with lots of breaking changes.

This also required an upgrade of askama.